+++ include/parseargs.h
@@ -38,6 +38,7 @@ struct ceccomp_args
   print_mode fmt_mode;
 
   bool quiet;
+  bool color;
   char *syscall_nr;
   uint64_t sys_args[6];
   uint64_t ip;