+++ src/ceccomp.c
@@ -1,5 +1,6 @@
 // this is for unit test
 #include "asm.h"
+#include "color.h"
 #include "disasm.h"
 #include "emu.h"
 #include "log/logger.h"
@@ -9,6 +10,7 @@
 #include "transfer.h"
 #include <argp.h>
 #include <assert.h>
+#include <stdbool.h>
 #include <stdint.h>
 #include <stdio.h>
 #include <string.h>
@@ -27,6 +29,7 @@ init (ceccomp_args *args)
   args->read_fp = stdin;
   args->fmt_mode = HEXLINE;
   args->quiet = false;
+  args->color = true;
   args->syscall_nr = (char *)ARG_INIT_VAL;
   args->sys_args[0] = ARG_INIT_VAL;
   args->sys_args[1] = ARG_INIT_VAL;
@@ -39,9 +42,13 @@ init (ceccomp_args *args)
 }
 
 static struct argp_option options[] = {
-  { "quiet", 'q', NULL, 0, NULL, 0 },  { "output", 'o', "OUTPUT", 0, NULL, 0 },
-  { "arch", 'a', "ARCH", 0, NULL, 0 }, { "pid", 'p', "PID", 0, NULL, 0 },
-  { "fmt", 'f', "FMT", 0, NULL, 0 },   { "help", 'h', NULL, 0, NULL, 0 },
+  { "quiet", 'q', NULL, 0, NULL, 0 },
+  { "no-color", 'c', NULL, 0, NULL, 0 },
+  { "output", 'o', "OUTPUT", 0, NULL, 0 },
+  { "arch", 'a', "ARCH", 0, NULL, 0 },
+  { "pid", 'p', "PID", 0, NULL, 0 },
+  { "fmt", 'f', "FMT", 0, NULL, 0 },
+  { "help", 'h', NULL, 0, NULL, 0 },
   { "usage", 'u', NULL, 0, NULL, 0 },
 };
 
@@ -62,6 +69,9 @@ main (int argc, char **argv)
   uint32_t program_start_idx = -1;
   if (args.mode == TRACE_PROG_MODE || args.mode == PROBE_MODE)
     program_start_idx = args.program_idx;
+
+  if (args.color)
+    disable_color ();
 
   switch (args.mode)
     {